mkfs.jffs2: Add sortfiles option.
Due to JFFS2 garbage collection and wear leveling, the kernel and cfe
files may be relocated. This causes problems for cfe.

This is a TEMPORARY fix that makes it possible to allocate specified
files to their own erase blocks thus making them less likely for
garbage collection (however eventually the files wil be moved so this
only postpones the problem).