mkfs.ubifs: Added support for LZMA compression
This is partly based on the OpenWRT patch with the same purpose but
it allows using LZMA as default compressor and it uses the same local
compression library as mkfs.jffs2 instead of an external library.
Also in this patch the algorithm is called "lzma" and not "xz" as in
the OpenWRT patch.